mturk/requirements.txt
boto3
botocore
psycopg2-binary